OpenAPI

+++ trillas-library/src/main/java/com/hypermedia/library/user/Token.java
@@ -2,7 +2,6 @@ package com.hypermedia.library.user;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
 
@@ -16,7 +15,7 @@ public class Token {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     private String token;
